pyproject.toml
[project]
name = "fastapi-money-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.10.10",
    "fastapi[standard]>=0.115.4",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
